Fitness Coach Agent/pyproject.toml
[project]
name = "exam-prepration-5-to"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.2.11",
    "rich>=14.1.0",
]
